+++ collectors/CRT.py
@@ -9,7 +9,7 @@ def init(domain):
 	print(colored("[*]-Searching CRT...", "yellow"))
 
 	parameters = {"q": "%.{0}".format(domain), "output": "json"}
-	headers = {"user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.10; rv:52.0) Gecko/20100101 Firefox/52.0", "content-type": "application/json"}
+	headers = {"user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:88.0) Gecko/20100101 Firefox/88.0", "content-type": "application/json"}
 
 	try:
 		response = requests.get("https://crt.sh/?", params=parameters, headers=headers)