+++ collectors/GoogleTransparency.py
@@ -21,7 +21,7 @@ def init(domain):
 	print(colored("[*]-Searching Google Transparency...", "yellow"))
 
 	baseURL = "https://www.google.com/transparencyreport/api/v3/httpsreport/ct/certsearch"
-	headers = {"user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.10; rv:52.0) Gecko/20100101 Firefox/52.0", "referrer": "https://transparencyreport.google.com/https/certificates"}
+	headers = {"user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:88.0) Gecko/20100101 Firefox/88.0", "referrer": "https://transparencyreport.google.com/https/certificates"}
 	token = ""
 
 	try: