+++ collectors/ProjectCrobat.py
@@ -9,7 +9,7 @@ def init(domain, ranges):
 	print(colored("[*]-Searching Project Crobat...", "yellow"))
 
 	url = "https://sonar.omnisint.io/subdomains/{0}".format(domain)
-	headers = {"user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.10; rv:52.0) Gecko/20100101 Firefox/52.0"}
+	headers = {"user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:88.0) Gecko/20100101 Firefox/88.0"}
 
 	try:
 		response = requests.get(url, headers=headers)