+++ collectors/WaybackMachine.py
@@ -10,7 +10,7 @@ def init(domain):
 	print(colored("[*]-Searching WaybackMachine...", "yellow"))
 
 	url = "http://web.archive.org/cdx/search/cdx?url=*.{0}&output=json&fl=original&collapse=urlkey".format(quote(domain))
-	headers = {"user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.10; rv:52.0) Gecko/20100101 Firefox/52.0"}
+	headers = {"user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:88.0) Gecko/20100101 Firefox/88.0"}
 
 	try:
 		response = requests.get(url, headers=headers)