+++ lepus.py
@@ -12,6 +12,7 @@ import collectors.DNSTrails
 import collectors.GoogleTransparency
 import collectors.HackerTarget
 import collectors.PassiveTotal
+import collectors.PDChaos
 import collectors.ProjectCrobat
 import collectors.ProjectSonar
 import collectors.Riddler
@@ -20,16 +21,18 @@ import collectors.Spyse
 import collectors.ThreatCrowd
 import collectors.VirusTotal
 import collectors.WaybackMachine
+import collectors.ZoomEye
 import submodules.Permutations
 import submodules.PortScan
 import submodules.ReverseLookups
 import submodules.TakeOver
+import submodules.Markov
 import utilities.DatabaseHelpers
 import utilities.MiscHelpers
 import utilities.ScanHelpers
 
 simplefilter("ignore")
-version = "3.3.2"
+version = "3.4.0"
 
 
 def printBanner():
@@ -55,6 +58,11 @@ if __name__ == "__main__":
 	parser.add_argument("--portscan", action="store_true", dest="portscan", help="scan resolved public IP addresses for open ports", default=False)
 	parser.add_argument("-p", "--ports", action="store", dest="ports", help="set of ports to be used by the portscan module [default is medium]", type=str)
 	parser.add_argument("--takeover", action="store_true", dest="takeover", help="check identified hosts for potential subdomain take-overs", default=False)
+	parser.add_argument("--markovify", action="store_true", dest="markovify", help="use markov chains to identify more subdomains", default=False)
+	parser.add_argument("-ms", "--markov-state", action="store", dest="markov_state", help="markov state size [default is 3]", type=int, default=3)
+	parser.add_argument("-ml", "--markov-length", action="store", dest="markov_length", help="max length of markov substitutions [default is 5]", type=int, default=5)
+	parser.add_argument("-mq", "--markov-quantity", action="store", dest="markov_quantity", help="max quantity of markov results per candidate length [default is 5]", type=int, default=5)
+	parser.add_argument("-f", "--flush", action="store_true", dest="doFlush", help="purge all records of the specified domain from the database", default=False)
 	parser.add_argument("-v", "--version", action="version", version="Lepus v{0}".format(version))
 	args = parser.parse_args()
 
@@ -63,11 +71,18 @@ if __name__ == "__main__":
 
 	printBanner()
 
+	print("{0} {1}\n".format(colored("\n[*]-Running against:", "yellow"), colored(args.domain, "cyan")))
+
+	db = utilities.DatabaseHelpers.init()
+	utilities.ScanHelpers.retrieveDNSRecords(db, args.domain)
+	old_resolved, old_unresolved, old_takeovers = utilities.MiscHelpers.loadOldFindings(db, args.domain)
+	utilities.MiscHelpers.purgeOldFindings(db, args.domain)
+
+	if args.doFlush:
+		print("{0} {1} {2}".format(colored("\n[*]-Flushed", "yellow"), colored(args.domain, "cyan"), colored("from the database", "yellow")))
+		exit(0)
+
 	try:
-		db = utilities.DatabaseHelpers.init()
-		utilities.ScanHelpers.retrieveDNSRecords(db, args.domain)
-		old_resolved, old_unresolved, old_takeovers = utilities.MiscHelpers.loadOldFindings(db, args.domain)
-		utilities.MiscHelpers.purgeOldFindings(db, args.domain)
 
 		if args.zoneTransfer:
 			zt_subdomains = utilities.ScanHelpers.zoneTransfer(db, args.domain)
@@ -88,6 +103,7 @@ if __name__ == "__main__":
 			collector_subdomains += collectors.GoogleTransparency.init(args.domain)
 			collector_subdomains += collectors.HackerTarget.init(args.domain)
 			collector_subdomains += collectors.PassiveTotal.init(args.domain)
+			collector_subdomains += collectors.PDChaos.init(args.domain)
 			collector_subdomains += collectors.ProjectCrobat.init(args.domain, args.ranges)
 			collector_subdomains += collectors.ProjectSonar.init(args.domain)
 			collector_subdomains += collectors.Riddler.init(args.domain)
@@ -96,6 +112,7 @@ if __name__ == "__main__":
 			collector_subdomains += collectors.ThreatCrowd.init(args.domain)
 			collector_subdomains += collectors.VirusTotal.init(args.domain)
 			collector_subdomains += collectors.WaybackMachine.init(args.domain)
+			collector_subdomains += collectors.ZoomEye.init(args.domain)
 
 		if args.wordlist:
 			wordlist_subdomains = utilities.MiscHelpers.loadWordlist(args.domain, args.wordlist)
@@ -124,6 +141,9 @@ if __name__ == "__main__":
 			if args.reverse:
 				submodules.ReverseLookups.init(db, args.domain, args.ranges, args.threads)
 
+			if args.markovify:
+				submodules.Markov.init(db, args.domain, args.markov_state, args.markov_length, args.markov_quantity, args.hideWildcards, args.threads)
+
 			utilities.ScanHelpers.massRDAP(db, args.domain, args.threads)
 
 			if args.portscan:
@@ -132,8 +152,9 @@ if __name__ == "__main__":
 			if args.takeover:
 				submodules.TakeOver.init(db, args.domain, old_takeovers, args.threads)
 
-		utilities.MiscHelpers.exportFindings(db, args.domain, old_resolved)
+		utilities.MiscHelpers.exportFindings(db, args.domain, old_resolved, False)
 
 	except KeyboardInterrupt:
-		print(colored("\n[*]-Received keyboard interrupt! Shutting down...\n", "red"))
+		print(colored("\n[*]-Received keyboard interrupt! Shutting down...", "red"))
+		utilities.MiscHelpers.exportFindings(db, args.domain, old_resolved, True)
 		exit(-1)